split into client and server
